Backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
requests
python-multipart
